refactor: move specv2 → steplibrary/specgen
Relocates the V2 inventory generator under steplibrary/ alongside the
read path, and renames the package to specgen to avoid the
generator.Generate stutter.
