+++ steplibrary/specgen/testdata/input/steps/hello-step/step-info.yml
